// src/app/enums/TodoEnums.ts
export enum TodoEnum {
  ADD_TODO_REQUEST = 'ADD_TODO_REQUEST',
  ADD_TODO_SUCCESS = 'ADD_TODO_SUCCESS',
  ADD_TODO_FAILURE = 'ADD_TODO_FAILURE',
  GET_TODO_REQUEST = 'GET_TODO_REQUEST',
  GET_TODO_SUCCESS = 'GET_TODO_SUCCESS',
  GET_TODO_FAILURE = 'GET_TODO_FAILURE',
  SET_TODO_DEFAULT = 'SET_TODO_DEFAULT',
  UPDATE_TODO_SUCCESS = 'UPDATE_TODO_SUCCESS',
  UPDATE_TODO_FAILURE = 'UPDATE_TODO_FAILURE',
  UPDATE_TODO_REQUEST = 'UPDATE_TODO_REQUEST',
}